Cleaned up the Rectangle's drawOutline method even more.
Added the GameObjectIterable, so that you can easily iterate through a list of gameobjects and only deal with those active ones.  It also allows you to easily re-use gameobjects (as a pool) and clean up any inactive objects when your count is too high.

I need to modify OrbitBehavior to simply accelerate towards the center of an object, and perhaps 'kick' the velocity tangent to the orbit to stabalize things.  Otherwise, using center-seeking accel should:
1) reduce update complexity and target seeking
2) make other effects equally prominent- by only seeking a single target and simply adjusting speed caps, other effects can touch velocity safely without having their values overwritten in the next frame.  This will need some careful tuning, so I'll probably test on the existing XNA version before porting here.