Changed world to interface, added level interface.
I'm going to pull out the orb-specific code, and package the engine into a jar so I can layer projects over it.  To that end, I'm trying to keep things from being too restrictive, but not so abstract that they have no behavior.
I'll need to think about how to handle collision checking with an abstract world- before, I could know the lists of a world and check them.  Perhaps in world initialization you'll push (key, list) pairs to the physics engine, and then have a dictionary of (key,key) collision types to check.