Started trying to add shaders, removed custom vec2 class.
Having some trouble with shaders, but the Vec2 removal went pretty well.
There are some minor problems with steering, as it looks like the rectangles jitter badly.  Their turning should be smoother, so I'll have to look into that.  It's odd that Vector2 functions return themselves, meaning every function mutates the values.  This is actually really good for readability, though, as it allows chaining (which is awesome!)  What that means is somewhere in my replacing Vec2, I probably forgot to copy a Vector2 before changing its value, which is why the rectangles are jittering so badly- their position/velocity is getting modified instead of being copied.